# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "To-Do API",
    "description": "API REST para gerenciamento de to-dos com compartilhamento entre usuários. Construída com Next.js + TypeScript + Supabase.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desenvolvimento"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido no login"
      }
    },
...
  },
  "paths": {
...
    "/api/todos/shared": {
      "get": {
        "summary": "Lista to-dos compartilhados comigo",
        "tags": ["Sharing"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": { "description": "Lista de to-dos compartilhados" },
          "401": { "description": "Não autenticado" }
        }
      }
    }
  }
}